feat: add confusable cluster merger implementation and integrate into label-based clustering
