+++ perception/autoware_euclidean_cluster/CMakeLists.txt
@@ -19,6 +19,7 @@ ament_auto_add_library(${PROJECT_NAME}_lib SHARED
   lib/euclidean_cluster.cpp
   lib/voxel_grid_based_euclidean_cluster.cpp
   lib/utils.cpp
+  lib/confusable_cluster_merger.cpp
 )
 
 target_link_libraries(${PROJECT_NAME}_lib
