+++ perception/autoware_euclidean_cluster/src/label_based_euclidean_cluster_node.hpp
@@ -14,6 +14,7 @@
 
 #pragma once
 
+#include "autoware/euclidean_cluster/confusable_cluster_merger.hpp"
 #include "autoware/euclidean_cluster/euclidean_cluster_interface.hpp"
 #include "autoware/euclidean_cluster/voxel_grid_based_euclidean_cluster.hpp"
 
@@ -39,17 +40,6 @@ namespace autoware::euclidean_cluster
 enum ShapePolicy : uint8_t {
   ALL_POLYGON = 0,
   LABEL_DEPEND = 1,
-};
-
-struct ConfusableLabelGroup
-{
-  std::vector<std::uint8_t> labels;
-  /// @brief True minimum point-to-point XY gap (m) below which two cross-label clusters may merge.
-  float cross_label_tolerance{};
-  /// @brief Maximum diameter (m) of a merged component's XY bounding circle; caps chaining into
-  ///        oversized blobs. Orientation-independent — clusters are partial views, so a per-axis
-  ///        (length/width) cap would presuppose a heading we cannot reliably estimate.
-  float max_merged_size{25.0F};
 };
 
 /// @brief ROS 2 node that performs euclidean clustering on semantic point clouds grouped by label.
